Resolve promise with errors instead of rejecting
Previously, the validate() function returned a promise which
would resolve() if there were no validation errors, and reject(errors)
if there were errors. However, when calling this validate method, we called it as the callback
of the "fields-validation-trigger" event.

When this promise is rejected because of validation errors, we
do not have a .catch method to deal with the rejection. This causes
it to get "stuck" and dies in the console. This is unwanted behavior.

The javascript engine will track such unhandled rejections and generate
global errors with the event "unhandledrejection".

This fills the console with "errors" which aren't actual execution errors.

The solution I had here was to update the code to always use resolve. If there
are errors, we resolve with the errors, if there are no errors, we resolve with
undefined/null.

This was the best solution I could find because:

1. Promise-based async validators work the same way. If there is an error,
resolve with the error, else resolve with null. This is evident in the following
code snippet from the abstractField.js validate function:
Promise.all(validators)
       .then(handleErrors)
       .catch((e) => { ... });

Using the same behaviour then is consistent with this existing pattern.

2. The returned promise was not being used anywhere expect in the tests. Therefore,
updating this was a no-brainer. Just update the code, and update the tests.

3. While it is debatable, rejecting a promise implies that something went wrong that
should not go wrong. When validating fields, getting a validation error is not in the
class of "something went wrong". A validators work is to find errors in supplied inputs,
and when it finds them, we know that the code is 100% functional. Nothing went wrong,
everything went as it should. Therefore, rejecting the promise feels like a bad pattern
because the code to find the errors just worked like it should and gave back the output
we want from it.